refactor: move libc module
