+++ pwnlib/libc/__init__.py
@@ -1,6 +1,6 @@
 """
 Module for convenient functions specific to different libcs.
 """
-from pwnlib.util.libc import glibc
+from pwnlib.libc import glibc
 
 __all__ = ['glibc']
